safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.6_Iteration_Statements/12.6.3_The_for_Statement/S12.6.3_A7_T1-expected.txt

CONSOLE MESSAGE: line 79: SyntaxError: Unexpected token ';'. Expected ')' to end a for-loop header.
S12.6.3_A7_T1

PASS Expected parsing failure

TEST COMPLETE
